requirements.txt

requests
feedparser
PyYAML
DingtalkChatbot
jinja2
